requirements.txt
scrapy
pyyaml
beautifulsoup4